Separated myAccount package
