+++ app/src/main/java/com/example/runtracker/myAccount/MyAccountItem.kt
@@ -1,4 +1,4 @@
-package com.example.runtracker
+package com.example.runtracker.myAccount
 
 class MyAccountItem(val item : String, val appWidget : String) {
 }
